Move ethereum:forkchoice into storage.
Resolves a cyclic dependency and moves protoarray into the same module as Store.
