+++ data/provider/build.gradle
@@ -3,7 +3,6 @@ dependencies {
 
     implementation project(':data:serializer')
     implementation project(':ethereum:core')
-    implementation project(':ethereum:forkchoice')
     implementation project(':ethereum:spec')
     implementation project(':ethereum:statetransition')
     implementation project(':infrastructure:async')
