--- ethereum/forkchoice/build.gradle
@@ -1,19 +1,0 @@
-dependencies {
-    implementation 'org.apache.tuweni:tuweni-ssz'
-    implementation project(':ethereum:spec')
-    implementation project(':infrastructure:async')
-    implementation project(':infrastructure:events')
-    implementation project(':infrastructure:exceptions')
-    implementation project(':infrastructure:logging')
-
-    testImplementation project(':infrastructure:crypto')
-    testImplementation testFixtures(project(':ethereum:core'))
-    testImplementation testFixtures(project(':ethereum:spec'))
-    testImplementation testFixtures(project(':storage'))
-
-    testFixturesApi 'org.apache.tuweni:tuweni-bytes'
-    testFixturesImplementation project(':ethereum:spec')
-    testFixturesImplementation testFixtures(project(':ethereum:spec'))
-    testFixturesImplementation project(':infrastructure:async')
-
-}
