+++ ethereum/weaksubjectivity/build.gradle
@@ -1,6 +1,5 @@
 dependencies {
     implementation project(':ethereum:core')
-    implementation project(':ethereum:forkchoice')
     implementation project(':ethereum:spec')
     implementation project(':infrastructure:async')
     implementation project(':infrastructure:logging')
