+++ fork-choice-tests/build.gradle
@@ -2,7 +2,6 @@ jar { enabled = false }
 
 dependencies {
     integrationTestImplementation project(':ethereum:core')
-    integrationTestImplementation project(':ethereum:forkchoice')
     integrationTestImplementation testFixtures(project(':ethereum:spec'))
     integrationTestImplementation project(':ethereum:statetransition')
     integrationTestImplementation project(':eth-tests')
