+++ networking/eth2/build.gradle
@@ -2,7 +2,6 @@ dependencies {
   api project(':networking:p2p')
   implementation project(':ethereum:core')
   implementation project(':ethereum:events')
-  implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:spec')
   implementation project(':ethereum:statetransition')
   implementation project(':infrastructure:async')
