+++ services/chainstorage/build.gradle
@@ -1,6 +1,5 @@
 dependencies {
   implementation project(':ethereum:core')
-  implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:pow:api')
   implementation project(':ethereum:spec')
