+++ settings.gradle
@@ -9,7 +9,6 @@ include 'ethereum:dataproviders'
 include 'ethereum:events'
 include 'ethereum:executionengine'
 include 'ethereum:executionlayer'
-include 'ethereum:forkchoice'
 include 'ethereum:networks'
 include 'ethereum:pow:api'
 include 'ethereum:statetransition'
