+++ storage/api/src/main/java/tech/pegasys/teku/storage/api/StoredBlockMetadata.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.ethereum.forkchoice;
+package tech.pegasys.teku.storage.api;
 
 import java.util.Objects;
 import java.util.Optional;
