+++ storage/build.gradle
@@ -2,7 +2,6 @@ dependencies {
   implementation project(':beacon:pow')
   implementation project(':ethereum:core')
   implementation project(':ethereum:dataproviders')
-  implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:pow:api')
   implementation project(':ethereum:spec')
   implementation project(':infrastructure:async')
@@ -34,9 +33,9 @@ dependencies {
 
   testImplementation testFixtures(project(':beacon:pow'))
   testImplementation testFixtures(project(':infrastructure:bls'))
+  testImplementation project(':infrastructure:crypto')
   testImplementation testFixtures(project(':infrastructure:metrics'))
   testImplementation testFixtures(project(':ethereum:core'))
-  testImplementation testFixtures(project(':ethereum:forkchoice'))
   testImplementation project(':ethereum:networks')
   testImplementation testFixtures(project(':ethereum:spec'))
   testImplementation testFixtures(project(':infrastructure:async'))
@@ -53,7 +52,6 @@ dependencies {
   testFixturesImplementation testFixtures(project(':infrastructure:bls'))
   testFixturesImplementation testFixtures(project(':infrastructure:metrics'))
   testFixturesImplementation testFixtures(project(':ethereum:core'))
-  testFixturesImplementation testFixtures(project(':ethereum:forkchoice'))
   testFixturesImplementation testFixtures(project(':ethereum:spec'))
   testFixturesImplementation project(':ethereum:networks')
   testFixturesImplementation testFixtures(project(':infrastructure:async'))
