+++ storage/src/main/java/tech/pegasys/teku/storage/protoarray/ProtoArrayBuilder.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.ethereum.forkchoice;
+package tech.pegasys.teku.storage.protoarray;
 
 import static com.google.common.base.Preconditions.checkNotNull;
