+++ storage/src/test/java/tech/pegasys/teku/storage/protoarray/AbstractBlockMetadataStoreTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.ethereum.forkchoice;
+package tech.pegasys.teku.storage.protoarray;
 
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toSet;
