+++ teku/build.gradle
@@ -12,7 +12,6 @@ dependencies {
   implementation project(':ethereum:core')
   implementation project(':ethereum:dataproviders')
   implementation project(':ethereum:executionengine')
-  implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:pow:api')
   implementation project(':ethereum:spec')
